+++ test/unit/package_admin/unit_test_run3.sh
@@ -21,4 +21,5 @@ echo "Unit test:  package_admin.py"
 /usr/bin/python3 ./test/unit/package_admin/rabbitmq_publish.py
 /usr/bin/python3 ./test/unit/package_admin/run_program.py
 /usr/bin/python3 ./test/unit/package_admin/main.py
+/usr/bin/python3 ./test/unit/package_admin/write_file.py